DIV-2657
Extracted out test utility classes in automated-tests module to reuse in new integration test required for draft API

+++ src/main/java/uk/gov/hmcts/reform/divorce/draftservice/service/DraftsService.java
@@ -61,7 +61,7 @@ public class DraftsService {
         DraftsResponse draftsResponse = draftsRetrievalService.getDraft(jwt,
                 userDetails.getId(),
                 encryptionKeyFactory.createEncryptionKey(userDetails.getId()));
-        if (draftsResponse.isDraft()) {
+        if (draftsResponse != null && draftsResponse.isDraft()) {
             draftStoreClient.deleteDraft(jwt, draftsResponse.getDraftId());
         }
     }
